requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Deprecated==1.2.13
Django==3.2.10
django-cors-headers==3.10.0
django-oauth-toolkit==1.6.1
djangorestframework==3.13.1
idna==3.3
jwcrypto==1.0
oauthlib==3.1.1
psycopg2==2.9.2
pycparser==2.21
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
wrapt==1.13.3
